UI, build, and deployment enhacements

+++ src/frontend/src/App.tsx
@@ -1,10 +1,11 @@
 import axios from "axios"
-import { For, createSignal } from "solid-js"
+import { createSignal } from "solid-js"
 import { encode } from 'gpt-tokenizer'
 import { makePersisted } from "@solid-primitives/storage"
-import { sample1, sample2, samplePrompt } from "./components"
+import { sample1, sample2, samplePrompt, uuidv4 } from "./components"
 import { Header } from "./components/Header"
 import { Information } from "./components/Information"
+import EmbeddingsArea from "./components/embeddingarea"
 
 const SplittingMethod = {
   SK: "SK",
@@ -27,18 +28,17 @@ const DefaultSettings: ISettings = {
   url: ""
 }
 
-const BASE_URI = "http://localhost:5096/"
+const BASE_URI = import.meta.env.VITE_BASE_URI;//"http://localhost:5096/"
 const URI_CHUNK = BASE_URI + "api/v1/content/split"
 const URI_LOAD = BASE_URI + "api/v1/content/load"
 //const URI_COMPLETION = BASE_URI + "api/v1/content/completion"
 const URI_RAG_QUERY = BASE_URI + "api/v1/rag/query"
 const URI_RAG_RESET_USER = BASE_URI + "api/v1/rag/reset/{userId}"
 
-
 function App() {
   const [settings, setSettings] = makePersisted(createSignal<ISettings>(DefaultSettings))
-  const [text, setFile1Text] = makePersisted(createSignal(''))
-  const [text2, setFile2Text] = makePersisted(createSignal(''))
+  const [contentFile1, setContentFile1] = makePersisted(createSignal(''))
+  const [contentFile2, setContentFile2] = makePersisted(createSignal(''))
   const [file1Tokens, setFile1Tokens] = createSignal(0)
   const [file2Tokens, setFile2Tokens] = createSignal(0)
   const [tokensContext, setTokensContext] = createSignal(0)
@@ -54,26 +54,15 @@ function App() {
   const [tab, setTab] = createSignal('chunk')
   const [userId, setUserId] = makePersisted(createSignal(uuidv4()))
 
-  function uuidv4() {
-    let userId = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'
-      .replace(/[xy]/g, function (c) {
-        const r = Math.random() * 16 | 0,
-          v = c == 'x' ? r : (r & 0x3 | 0x8);
-        return v.toString(16);
-      });
-    userId = userId.replace(/-/g, '')
-    return "user_" + userId.substring(userId.length - 12)
-  }
-
   const getTokenCountAfterTyping = (value: string, control: string) => {
     if (control === "chunk") {
-      setFile1Text(value)
+      setContentFile1(value)
       setFile1Tokens((encode(value)).length);
-      setFile2Tokens((encode(text2())).length);
+      setFile2Tokens((encode(contentFile2())).length);
     }
     if (control === "chunk2") {
-      setFile2Text(value)
-      setFile1Tokens((encode(text())).length);
+      setContentFile2(value)
+      setFile1Tokens((encode(contentFile1())).length);
       setFile2Tokens((encode(value)).length);
     }
     if (control === "context") {
@@ -87,44 +76,44 @@ function App() {
   }
 
   const UpdateTokenCounts = () => {
-    setFile1Tokens((encode(text())).length);
-    setFile2Tokens((encode(text2())).length);
+    setFile1Tokens((encode(contentFile1())).length);
+    setFile2Tokens((encode(contentFile2())).length);
     setTokensPrompt((encode(settings().prompt + context())).length);
     setTokensContext(encode(context()).length);
     setTokensCompletion(encode(completion()).length);
   }
 
-  const LoadContext = () => {
-    const totalChunks = parseInt(settings().chunks)
-    const chunks = allMemories()
-    if (totalChunks > 0 && allMemories().length > 0) {
-      let chunkText = ""
-      for (let i = 0; i < totalChunks; i++) {
-        if (i == chunks.length) break;
-        chunkText += chunks[i].text + "\n\n"
-      }
-      setContext(chunkText)
-      UpdateTokenCounts()
-    }
-  }
+  // const LoadContext = () => {
+  //   const totalChunks = parseInt(settings().chunks)
+  //   const chunks = allMemories()
+  //   if (totalChunks > 0 && allMemories().length > 0) {
+  //     let chunkText = ""
+  //     for (let i = 0; i < totalChunks; i++) {
+  //       if (i == chunks.length) break;
+  //       chunkText += chunks[i].text + "\n\n"
+  //     }
+  //     setContext(chunkText)
+  //     UpdateTokenCounts()
+  //   }
+  // }
 
   const ProcessChunks = async () => {
     if (processing()) return
     setProcessing(true)
     setAllMemories([])
     setUsedMemories([])
-    getTokenCountAfterTyping(text(), "chunk")
+    getTokenCountAfterTyping(contentFile1(), "chunk")
     const maxTokensPerParagraph = settings().method == SplittingMethod.Paragraph ? parseInt(settings().wordCount) : parseInt(settings().maxTokensPerLine)
     const payload = {
       userName: userId(),
       files: [
         {
           fileName: "file1",
-          content: text()
+          content: contentFile1()
         },
         {
           fileName: "file2",
-          content: text2()
+          content: contentFile2()
         }
       ],
       maxTokensPerLine: parseInt(settings().maxTokensPerLine),
@@ -147,10 +136,10 @@ function App() {
     }
   }
 
-  const LoadAndProcess = async () => {
+  const LoadSampleFilesAndProcess = async () => {
     if (processing()) return
-    setFile1Text(sample1)
-    setFile2Text(sample2)
+    setContentFile1(sample1)
+    setContentFile2(sample2)
     await ProcessChunks()
   }
 
@@ -167,7 +156,7 @@ function App() {
     try {
       const resp = await axios.post(URI_LOAD, payload)
       const content = resp.data.content;
-      setFile1Text(content)
+      setContentFile1(content)
       UpdateTokenCounts()
     } catch (err) {
       console.error(err)
@@ -204,11 +193,13 @@ function App() {
       console.error(err)
     } finally {
       setProcessing(false)
-      setPromptButtonLabel("Submit")
+      setPromptButtonLabel("Process")
     }
   }
 
-  const Reset = async () => {
+  const ResetSessionAndUser = async () => {
+    if (processing()) return
+    setProcessing(true)
     try {
       if (userId())
         await axios.delete(URI_RAG_RESET_USER.replace("{userId}", userId()))
@@ -216,8 +207,8 @@ function App() {
       console.error(err)
     }
     setSettings(DefaultSettings)
-    setFile1Text('')
-    setFile2Text('')
+    setContentFile1('')
+    setContentFile2('')
     setFile1Tokens(0)
     setFile2Tokens(0)
     setTokensContext(0)
@@ -231,6 +222,7 @@ function App() {
     setPromptButtonLabel("Process")
     setUserId(uuidv4())
     setTab("chunk")
+    setProcessing(false)
   }
 
   return (
@@ -262,14 +254,14 @@ function App() {
             <span title="Total Tokens" class='px-2 bg-blue-900 text-white rounded-xl'>{tokensPrompt()}</span>
           </button>
           <button
-            onClick={() => setTab("context")}
-            class={`p-2 hover:underline hover:text-black ${tab() == "context" ? "border-black border-2 underline text-black" : "border"} font-semibold`}>Context&nbsp;
-            <span title="Total Tokens" class='px-2 bg-blue-900 text-white rounded-xl'>{tokensContext()}</span>
-          </button>
-          <button
             onClick={() => setTab("completion")}
             class={`p-2 hover:underline hover:text-black ${tab() == "completion" ? "border-black border-2 underline text-black" : "border"} font-semibold`}>Completion&nbsp;
             <span title="Total Tokens" class='px-2 bg-blue-900 text-white rounded-xl'>{tokensCompletion()}</span>
+          </button>
+          <button
+            onClick={() => setTab("context")}
+            class={`p-2 hover:underline hover:text-black ${tab() == "context" ? "border-black border-2 underline text-black" : "border"} font-semibold`}>Context&nbsp;
+            <span title="Total Tokens" class='px-2 bg-blue-900 text-white rounded-xl'>{tokensContext()}</span>
           </button>
         </nav>
       </div>
@@ -344,7 +336,7 @@ function App() {
                   <label class='font-bold uppercase'>File 1: <span class="font-bold bg-blue-800 rounded-xl text-white p-1">{file1Tokens()}</span></label>
                   <textarea
                     class="border border-black p-2 round-lg"
-                    value={text()}
+                    value={contentFile1()}
                     onInput={(e) => { getTokenCountAfterTyping(e.currentTarget.value, "chunk") }}
                     rows={25}>
                   </textarea>
@@ -352,7 +344,7 @@ function App() {
                 <div class="flex flex-col w-1/2 p-1">
                   <label class='font-bold uppercase'>File 2: <span class="font-bold bg-blue-800 rounded-xl text-white p-1">{file2Tokens()}</span></label>
                   <textarea
-                    value={text2()}
+                    value={contentFile2()}
                     onInput={(e) => { getTokenCountAfterTyping(e.currentTarget.value, "chunk2") }}
                     class="border border-black p-2 round-lg w-full"
                     rows={25}>
@@ -364,7 +356,7 @@ function App() {
                   onClick={ProcessChunks}
                   class="w-24 p-2 bg-blue-900 hover:bg-blue-900 text-white font-semibold">Embed</button>
                 <button
-                  onClick={LoadAndProcess}
+                  onClick={LoadSampleFilesAndProcess}
                   class="p-2 bg-green-800 hover:bg-green-700 text-white font-semibold">Load Samples & Embed</button>
               </div>
               <Information />
@@ -379,9 +371,9 @@ function App() {
                 onInput={(e) => getTokenCountAfterTyping(e.currentTarget.value, "context")}
                 rows={20}>
               </textarea>
-              <button class="p-2 w-24 bg-blue-900 text-white"
+              {/* <button class="p-2 w-24 bg-blue-900 text-white"
                 onclick={LoadContext}
-              >Load</button>
+              >Load</button> */}
             </div>
           </section>
           <section hidden={tab() !== "prompt"}>
@@ -446,53 +438,17 @@ function App() {
           </section>
         </div>
         <div class="w-full md:w-1/4 px-2 bg-blue-100">
-          <div class="space-x-2 bg-blue-900 text-white p-1">
-            <label class="uppercase text-sm">All Embeddings: </label> <span title="Total Chunks" class='px-2 bg-slate-800 text-white rounded-xl'>{allMemories().length}</span> -
-            <span title="Total Tokens" class='px-2 bg-slate-800 text-white rounded-xl'>{allMemories().reduce((acc, chunk) => acc + chunk.tokenCount, 0)}</span>
-          </div>
-          <div class="overflow-auto">
-            <For each={allMemories()}>
-              {(chunk) => (
-                <div class="flex flex-col mb-2 rounded-lg p-2 space-y-3 border-2 border-slate-300 hover:border-2 hover:border-slate-800 shadow">
-                  <div>
-                    <span class="text-sm font-bold uppercase">Embedding: </span>{chunk.chunkId} <span class="text-sm font-bold uppercase">Tokens: </span>{chunk.tokenCount}
-                  </div>
-                  <hr class="border-black" />
-                  {/* <span class="">{chunk.text}</span> */}
-                  <textarea class="bg-slate-200 p-2" rows={10} value={chunk.text} readOnly></textarea>
-                  <textarea class="bg-yellow-200 p-2" rows={2} value={chunk.embedding} readOnly></textarea>
-                </div>
-              )}
-            </For>
-          </div>
+          <EmbeddingsArea title="All Embeddings" memories={allMemories()} bg_color="bg-blue-900" />
         </div>
         <div class="w-full md:w-1/4 px-2 bg-green-100">
-          <div class="space-x-2 bg-green-800 text-white p-1">
-            <label class="uppercase text-sm">Used Embeddings: </label> <span title="Total Chunks" class='px-2 bg-slate-800 text-white rounded-xl'>{usedMemories().length}</span> -
-            <span title="Total Tokens" class='px-2 bg-slate-800 text-white rounded-xl'>{usedMemories().reduce((acc, chunk) => acc + chunk.tokenCount, 0)}</span>
-          </div>
-          <div class="overflow-auto">
-            <For each={usedMemories()}>
-              {(chunk) => (
-                <div class="flex flex-col mb-2 rounded-lg p-2 space-y-3 border-2 border-slate-300 hover:border-2 hover:border-slate-800 shadow">
-                  <div>
-                    <span class="text-sm font-bold uppercase">Embedding: </span>{chunk.chunkId} <span class="text-sm font-bold uppercase">Tokens: </span>{chunk.tokenCount}
-                  </div>
-                  <hr class="border-black" />
-                  {/* <span class="">{chunk.text}</span> */}
-                  <textarea class="bg-slate-200 p-2" rows={10} value={chunk.text} readOnly></textarea>
-                  <textarea class="bg-yellow-200 p-2" rows={2} value={chunk.embedding} readOnly></textarea>
-                </div>
-              )}
-            </For>
-          </div>
+          <EmbeddingsArea title="Used Embeddings" memories={usedMemories()} bg_color="bg-green-800" />
         </div>
       </div>
       <section class={"flex h-[35px] space-x-2 items-center px-2 " + (processing() ? "bg-red-600" : "bg-slate-800")}>
         <label class="text-white">ID:</label>
         <input class="px-1 text-black text-sm h-[24px] bg-slate-300" readOnly type="text" value={userId()} />
         <button class="bg-red-700 text-white p-1 hover:bg-red-600 hover:text-white font-semibold rounded-md"
-          onclick={Reset}
+          onclick={ResetSessionAndUser}
         >Reset</button>
       </section>
     </>
